requirements.txt
cffi==1.15.1
click==8.1.3
cryptography==38.0.1
Flask==2.2.2
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
peewee==3.15.3
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.5
six==1.16.0
Werkzeug==2.2.2
zipp==3.10.0
